python/requirements.txt
python-dotenv>=1.0.1
arcadepy>=1.7.0
openai>=1.40.0
openai-agents>=0.2.7
agents-arcade>=0.0.4; python_version >= "3.10"
eval_type_backport>=0.2.2; python_version < "3.10"

